requirements.txt
streamlit
PyMuPDF
python-docx
spacy
scikit-learn
openai
python-dotenv
